WIP: Gateway Crash
This commit aims to solve the issue described below:
The dashboard is not able to receive the message from the steering.

That is, the dashboard updates the testCount and increases success counter for the corresponding can message object but all of it's data is skewed. (It was all zero when I tested, so was the third message from VCU.)

Its arriving rate is rather higher than expected and seems like it is somehow reading the third message from VCU, which ought to be read from the other CAN node.

My theory is that something has crashed during the initializing process and the main suspect is that the iterating sequence of the for loops are somehow unexpectedly is altered during the compiler optimization.